refactor(desktop): organize cloud product hooks
