+++ desktop/src/hooks/access/cloud/use-cloud-credentials.ts
@@ -1,7 +1,7 @@
 import { useQuery } from "@tanstack/react-query";
 import { ProliferateClientError } from "@/lib/access/cloud/client";
 import { listCloudCredentialStatuses } from "@/lib/access/cloud/credentials";
-import { useCloudAvailabilityState } from "@/hooks/cloud/use-cloud-availability-state";
+import { useCloudAvailabilityState } from "@/hooks/cloud/derived/use-cloud-availability-state";
 import { useTauriCredentialsActions } from "@/hooks/access/tauri/use-credentials-actions";
 import { cloudCredentialsKey } from "@/hooks/access/cloud/query-keys";
 import type { CloudCredentialStatus } from "@/lib/domain/cloud/credentials";
