+++ desktop/src/hooks/cloud/derived/use-cloud-availability-state.ts
