+++ desktop/src/hooks/cloud/derived/use-cloud-repo-setup-suggestions.ts
