+++ desktop/src/hooks/cloud/facade/use-cloud-billing.ts
@@ -8,7 +8,7 @@ import {
 } from "@/hooks/access/cloud/use-cloud-billing";
 import { useTauriShellActions } from "@/hooks/access/tauri/use-shell-actions";
 import { useAppCapabilities } from "@/hooks/capabilities/derived/use-app-capabilities";
-import { useCloudAvailabilityState } from "@/hooks/cloud/use-cloud-availability-state";
+import { useCloudAvailabilityState } from "@/hooks/cloud/derived/use-cloud-availability-state";
 import { useAuthStore } from "@/stores/auth/auth-store";
 
 function billingOwnerKey(owner?: CloudOwnerSelection) {
