+++ desktop/src/hooks/workspaces/use-workspace-sidebar-actions.ts
@@ -3,7 +3,7 @@ import type { WorkspacePurgeResponse, WorkspaceRetireResponse } from "@anyharnes
 import { useToastStore } from "@/stores/toast/toast-store";
 import { APP_ROUTES } from "@/config/app-routes";
 import { useWorkspaceMobilityState } from "@/hooks/workspaces/mobility/use-workspace-mobility-state";
-import { useCreateCloudWorkspace } from "@/hooks/cloud/use-create-cloud-workspace";
+import { useCreateCloudWorkspace } from "@/hooks/cloud/workflows/use-create-cloud-workspace";
 import type { CloudWorkspaceRepoTarget } from "@/lib/domain/workspaces/cloud/cloud-workspace-creation";
 import type { SidebarIndicatorAction } from "@/lib/domain/workspaces/sidebar/sidebar";
 import { useWorkspaceEntryActions } from "./use-workspace-entry-actions";
